Major change in sources tree.
implementation of arm is on work it is a failed branch so yeah src/arm/ark is on work.better not to be touched with.and major modification is done in make file readme file will be updated to hold the changed listing. and added a build script for easy build for those who doesnt know.